refactor(backend-game-adapter-typeorm): move GameActionDb modules
