+++ packages/backend/apps/game/backend-game-adapter-typeorm/src/gameActions/adapter/typeorm/models/GameActionDb.ts
@@ -8,7 +8,7 @@ import {
   RelationId,
 } from 'typeorm';
 
-import { GameDb } from './GameDb';
+import { GameDb } from '../../../../games/adapter/typeorm/models/GameDb';
 
 @Entity({
   name: 'GameAction',
