+++ packages/backend/apps/game/backend-game-adapter-typeorm/src/gameActions/adapter/typeorm/models/GameActionKindDb.ts
@@ -1,4 +1,5 @@
 export enum GameActionKindDb {
+  draw = 'draw',
   passTurn = 'passTurn',
   playCards = 'playCards',
 }
