+++ packages/backend/apps/game/backend-game-domain/src/gameActions/domain/valueObjects/DrawGameAction.ts
@@ -3,5 +3,5 @@ import { BaseGameAction } from './BaseGameAction';
 import { GameActionKind } from './GameActionKind';
 
 export interface DrawGameAction extends BaseGameAction<GameActionKind.draw> {
-  readonly draw: Card | null;
+  readonly draw: Card[];
 }
